refactor: extract validation and mutation catalogs
